pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "epistemic-dialectic-engine"
version = "0.1.0"
description = "LLM pipeline operationalizing the Mindset Elevation Framework"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "streamlit>=1.30",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = ["slow: marks tests that hit real APIs"]
